Handle missing production and consumption OBIS codes
This PR adds logic for handling missing production and consumption OBIS codes.

Currently the client returns 0 if the registers:

    "01-00:01.08.00*ff"
    "01-00:02.08.00*ff"

are not present. This behavior is correct, as a user might think that the value is 0, although it is not present.

Therefore we add a return value of None, when missing.

Additionally tests are added.